api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
supabase==2.3.4
psycopg2-binary==2.9.9
gTTS==2.5.1
python-multipart==0.0.6
websockets==12.0
apscheduler==3.10.4
pyjwt==2.8.0
requests==2.31.0
